+++ plugin/src/main/java/io/github/mfederczuk/gradle/plugin/ktlint/KtlintPluginExtension.kt
@@ -48,9 +48,6 @@ public interface KtlintPluginExtension {
 	 *
 	 * Adds the flag `--code-style=<codeStyle>` to the ktlint invocation.
 	 *
-	 * This property requires that the [version] property be set to `0.49.0` or above, otherwise configuration will
-	 * fail.
-	 *
 	 * The default is no explicit code style / whatever ktlint uses as the default.
 	 *
 	 * @see KtlintOfficial
@@ -60,24 +57,6 @@ public interface KtlintPluginExtension {
 	public val codeStyle: Property<String>
 
 	// endregion
-
-	/**
-	 * Whether or not this project is an android project.
-	 *
-	 * If set to `true`, prior to ktlint version `0.49.0`, then the flag `--android` will be added to the `ktlint`
-	 * invocation. For ktlint version `0.49.0` and onwards, the flag `--code-style=android_studio` is added instead.
-	 *
-	 * This and the [codeStyle] property are mutually exclusive; If both are set, the configuration will fail.
-	 *
-	 * The default value is `false`.
-	 */
-	@Deprecated(
-		message = "Since ktlint version 0.49.0, the flag '--android' is deprecated in favor of " +
-			"'--code-style=android_studio', though it is recommended to instead set " +
-			"the EditorConfig property 'ktlint_code_style=android_studio' all together. " +
-			"If the configured version is before 0.49.0, then this warning can be ignored",
-	)
-	public val android: Property<Boolean>
 
 	/**
 	 * The maximum number of errors to show.